ELBScheme supports internal or internet-facing
I'm not sure this is totally correct, many months ago I had to find out the hard way that `internet-facing` fixed my problem.  However I just spun up a default app and it appears to set it to `public`

```
                {
                    "Namespace": "aws:ec2:vpc",
                    "OptionName": "ELBScheme",
                    "Value": "public"
                },
```

So I now don't know if it should be `internet-facing` or `public` .. but it definitely accepts more options than just `internal`.

I also had issues with simply removing the option entirely in the past, because if I wanted to swap from an `internal` to an internet facing one, removing the option didn't seem to default it back to public.