Note that Procfile should not include UTF-8 BOM marker
I'm suggesting a change to the documentation to indicate that if the `Procfile` (at least, untested others) is encoded as `UTF-8 with BOM marker` (byte order mark) the file is unrecognized by EB during deployment, with this error:

```
[ERROR] There is no proc command in proc file. Aborting the deployment.
```

This led to some confusion, as the file definitely existed and matched what the documentation provided as an example. Unfortunately, Visual Studio (Windows) for me included a BOM marker in the initial save for the `Procfile` leading to this issue.